scraper/requirements.txt
mysql-connector-python
selenium